requirements.txt
Flask==0.11.1
attrdict==2.0.0
eventlet==0.19.0
python-socketio==1.6.1
numpy==1.13.1
Pillow==2.2.1
scipy==0.19.1
keras==2.0.9
tensorflow-gpu==1.3.0
h5py==2.6.0
